Another, no, scratch that: THE tutorial improvement for tiles. Using horrific amounts of #ifdef USE_TILE etc, the tutorial texts now give different information for tiles (mouse clicks) and non-tiles (key commands).
Unfortunately, while I am confident this is a vast improvement
the new messages for tiles tend to be rather wordy because
"clicking on the item with your left mouse button" is much
verbose than "pressing 'w'". That, and most of the time I want
to print the tile information *in addition* to the general 
commands. So, it would be helpful if someone else could go
over the texts at one point - there's no need to hurry, though.

Also a bit of general improvement, which is always nice. I'd
somehow forgotten to include a tutorial trigger for mutations,
and of course the first potion I tried was mutation. :)


git-svn-id: https://crawl-ref.svn.sourceforge.net/svnroot/crawl-ref/trunk@3941 c06c8d41-db1a-0410-9941-cceddc491573