Add DEBUG_PANE_BOUNDS, which draws little bits at the upper-left and right of each pane.  The HUD bits are overwritten by the hud display, but as the comment notes:
    // Doesn't work for HUD because print_stats() overwrites it.
    // To debug HUD, add viewwindow(false,false) at end of _prep_input.

Add checks that no panes extend beyond the term size, or all the way
to the term bottom. (Bottom line is not available unless we can keep
the screen from scrolling when we write to it)

Fix a bunch of off-by-one errors.  Using a point vs vector metaphor
seems to flush them out nicely (hooray for affine math!)

There are still problems with the monster list resizing too big and
squishing the main map view -- didn't address any of that.


git-svn-id: https://crawl-ref.svn.sourceforge.net/svnroot/crawl-ref/trunk@3959 c06c8d41-db1a-0410-9941-cceddc491573