Refactor the layout logic for the inline and mlist-col layouts into a couple classes.  Hopefully it will be much easier to tune the layouts now.
Fixed problem I introduced earlier today that didn't let message pane
use the bottom-most line.  Stole a couple cols from the HUD and one
from the HUD gutter to potentially give to the mlist or to the main
view.  Stole a row from the message area so even 80x24 players get
get one line of monster list

Made the mlist-col layout not so unfair to the main view, and not
truncate the mlist width so much.  It used to pop in if you had an
81-col display; now the display needs to be larger (~95) for it to be
used, but it is a lot more functional.



git-svn-id: https://crawl-ref.svn.sourceforge.net/svnroot/crawl-ref/trunk@3972 c06c8d41-db1a-0410-9941-cceddc491573