Modify manual and readme to really list the new doc names, and key settings. Add vampires to manual (still needs reference to aptitudes etc.!) and reorder classes according to the new job order.
Re-introduce 't' as a synonym for 'f' in targetting, but don't 
document it anymore. (It's only in convenience to older players, 
anyway.) Also, change default wand inscription from !z to !Z, but
use a hack to check for either.

Oh, and restrict Blood Evaporate effect to Stinking cloud, Fire
and Steam (the latter two like !Rage).


git-svn-id: https://crawl-ref.svn.sourceforge.net/svnroot/crawl-ref/trunk@3981 c06c8d41-db1a-0410-9941-cceddc491573