Fix 1869850: Targeting now respects mimics' (un)known status correctly,              so that * and / work for unknown ones, and + and - for known 	     ones. Also, mimics become known if you hit them since the 	     resulting message ("moderately damaged" or "looks slightly 	     unstable") will give that away anyway.
Implement the first part of FR 1903048: Added a range parameter to
direction(), so that when targeting for weapons of reaching we can check
whether a monster is within the range of 2 squares (grid_distance).
Does not handle submerged monsters.
Eventually, this could be used to also limit max range for certain spells
(Freeze, I'm looking at you!)


git-svn-id: https://crawl-ref.svn.sourceforge.net/svnroot/crawl-ref/trunk@3995 c06c8d41-db1a-0410-9941-cceddc491573