Allow vampires to mutate when Satiated or higher, and make non-physical mutations only apply at these hunger levels, too.
Differentiated mutations into physical (affecting a character's
appearance and stuff: scales, hooves, wings, ...) and internal 
(resistances etc.). I guess this is what people usually refer to as 
"cosmetic", which I think is an inaccurate description since it makes 
it sound like they were completely superficial and had no real effect.

Here's a list of arbitrarily chosen "physical" mutations:
 * tough skin, all scales, and fur
 * strong/clever/agile, and weak/dopey/clumsy
 * deformed
 * strong but stiff, and flexible but weak
 * frail, and robust
 * claws, fangs, hooves, talons, and horns
 * stinger, wings
 * blue/green marks

Mutations currently not applying are still listed on the 'A' screen,
though in darkgrey and in (brackets).


git-svn-id: https://crawl-ref.svn.sourceforge.net/svnroot/crawl-ref/trunk@4246 c06c8d41-db1a-0410-9941-cceddc491573