Fix 1943303: Prevention of barding generation in monsters equipment.
And while I was at it... improved monster pickup handling.
* Ranged spellcasters won't pick up missiles anymore (except throwing nets).
* Allow monsters with behaviour != BEH_WANDER to pick up items again.
  (Restricted to scrolls, potions, wands and weapons, since these could be 
  useful in a fight.)
* Restrict monster pickup to items they may actually use (3 types of
  scrolls, six types of potions), though they may still pick up all types 
  of wands, if only to annoy the player.

As before, pickup only takes place if a monster happens to be on a square 
with items and only the top items of stashes are checked - monsters won't
make a beeline for items or search stashes for goodies.
Also, neutrals and friendlies still won't pick up anything at all; this might 
have to be changed. 

Unrelated: fixed an ASSERT for divine retribution during penance.


git-svn-id: https://crawl-ref.svn.sourceforge.net/svnroot/crawl-ref/trunk@4255 c06c8d41-db1a-0410-9941-cceddc491573