Fix tiles compile (in a hacky way, since I don't know what the mp/hp bar change was really about).
Modify mutations screen for Vampires to allow a toggle to a second page
that lists all those resistances and stuff depending on their blood
level. The screen (designed by David) is really neat, but the underlying 
could stand to be improved, and it might be too large, as well.

Also restrict Sublimation of Blood to Vampires that are at least Full,
and it also makes them a bit more thirsty (since they presumably lose
blood when using it). Dispel Undead only does half damage to Vampires at
Alive, and 66% damage at Full or Very Full.


git-svn-id: https://crawl-ref.svn.sourceforge.net/svnroot/crawl-ref/trunk@4325 c06c8d41-db1a-0410-9941-cceddc491573