Two changes:
1.) If Beogh can't find a nearby follower in LOS to bless, try again 
without the LOS restriction (but still nearby). If this also fails, recall 
a small amount of followers (1 + rnd(4) + rnd(4)) on the level. I don't
think there's much harm in Beogh actively doing something you already
have as an invocation as long as it doesn't make the invocation
superfluous (and this doesn't). 
In this case, the chance of this happening should probably be lowered, 
and actual reinforcement should be rarer still.

2.) Identify magical staves if you have at least 4 skill levels in the
corresponding spell school when wielding it, or gain the 4th level while
wielding it. (I decided on 4 because that is what most mages start out 
with in their special school, and it seems a good enough treshold.) 
Prior to that, there's still the random identification chance at casting
an appropriate spell, as before.


git-svn-id: https://crawl-ref.svn.sourceforge.net/svnroot/crawl-ref/trunk@4639 c06c8d41-db1a-0410-9941-cceddc491573