Fix 1941612: Zapping wands known to be empty shouldn't cost a turn, but              empty wands with unknown charges should cost a turn.
Experimentally add a "Really fire through friendly creature?" prompt
when a beam tracer passes through a friend. This currently uses the 
existing monster tracer and probably has huge problems because of this.
In any case, it appears to only work sometimes, though consistently for
wands in my testing, possibly because the range is fixed (?) when
compared to spells. (This is part of FR 1962548.)


git-svn-id: https://crawl-ref.svn.sourceforge.net/svnroot/crawl-ref/trunk@5171 c06c8d41-db1a-0410-9941-cceddc491573