Experimentally dump the new overview screen rather than the old one. I'm actually not entirely convinced this is the way to go. a) The new one is more informative, but the dump already contains    all that extra information in a much more detail. b) The new one is much more concise, but that's something that's not    necessary in the dump.
On the plus side, new players doesn't have to get used to two different
designs. Anyway, the old dump is still there, and we could use the
dump options to let the player choose which one to use.

For now, let's try whether it's possible to get used to this one.


git-svn-id: https://crawl-ref.svn.sourceforge.net/svnroot/crawl-ref/trunk@5225 c06c8d41-db1a-0410-9941-cceddc491573