When advising the player on resting, note that it's better to move back to an area where you've explored and killed off all the monsters, since if you rest on the edge of the explored areas you're more likely to be interrupted by a wandering monster.
On TUT_RETREAT_CASTER, if the player has wands that might be useful in
fighting, comment on that.

When first seeing an altar, tell conjurers that the best gods to
worship are Vehumet and Sif Muna.

When gaining your first god-granted ability, if the ability is passive
tell the player to use '^' instead of 'a' to examine the ability.


git-svn-id: https://crawl-ref.svn.sourceforge.net/svnroot/crawl-ref/trunk@5287 c06c8d41-db1a-0410-9941-cceddc491573