Fix a recent bug of mine that blocked the friendly fire prompts. Fix 1943573: Weird message order when attacking charmed monsters. Fix 1826896: NOTE_ID could have line breaks in mid-word
Modify the "This feature has more information" hint to mention
mouseclicking for Tiles players rather than 'v' like for purists.

Oh, and when trying to look into BR 1952908 (buggy trap in the Abyss)
I found that traps can be generated in the Abyss after all. I don't 
know when that was changed but the last time I checked I remember
seeing something like "if (level == ABYSS) return false" in a
"Can there be traps here?" check. Is my memory playing tricks on me
or was that really changed, and if so, was it intentional?


git-svn-id: https://crawl-ref.svn.sourceforge.net/svnroot/crawl-ref/trunk@5445 c06c8d41-db1a-0410-9941-cceddc491573