Revert the previous int -> unsigned in changes, as they don't match some of the code elsewhere.
Weird bits:

monsters::foe and player::pet_target are used interchangeably, but the 
former is an unsigned int, while the latter is an unsigned short.  Both 
of these can be set using the return value of monster_index(), which is 
a signed int.  Most of the functions that use these expect signed ints, 
as well.  ???


git-svn-id: https://crawl-ref.svn.sourceforge.net/svnroot/crawl-ref/trunk@5459 c06c8d41-db1a-0410-9941-cceddc491573