Add a line on good gods disapproving of certain items to their description (evil_item). Same for evil_eating.
In another step on the way to monster pathfinding, take the
shortest path and extract a vector of waypoints out of it.

When experimenting with ways to do this I noticed that grid_see_grid 
is not symmetrical (A may see B but not vice versa); I'm not sure 
what effects that could have. It won't directly affect the player as 
the checks for monster sees player and player sees monster both use 
the player LoS, but it could have an effect on friendly monsters 
fighting enemy ones, I guess.

Also, I don't think num_feats_between needs the shortest beam
available (called with false now). In fact, that seemed to hurt
visibility a bit, probably because of attempting to take vision
obstructing shortcuts. If this reasoning is wrong, please speak up
and/or correct it. (I sure hope not because the shortest beam 
calculation has some more overhead that can be avoided this way.)


git-svn-id: https://crawl-ref.svn.sourceforge.net/svnroot/crawl-ref/trunk@5501 c06c8d41-db1a-0410-9941-cceddc491573