Enable pathfinding for patrolling monsters who lost their patrol point. Fixed patrol definitions in maps not actually working.
Tested both by teleporting the minotaur in a labyrinth. About 50 turns
later it arrived back in its lair. :)

Since the minotaur proved to have a high magic resistance, it had plenty
of time to loot the stash at the labyrinth entrance once I'd woken it 
before I finally managed to teleport it away (in wizmode you can force 
monsters not noticing you).

This resulted in 
  "A minotaur, wielding the demon trident "Suyn Oma", and wearing a heavily
   runed ring mail."   o_O

It had also picked up a scroll and a wand. Ouch... (but cool!)


git-svn-id: https://crawl-ref.svn.sourceforge.net/svnroot/crawl-ref/trunk@5536 c06c8d41-db1a-0410-9941-cceddc491573