Nerf damage of the bat transformation. Reduced from   2 + skill(unarmed_combat)/3    to   2 + rnd(skill(unarmed_combat)/5 + 1)
Also replace the menu_colour_prefix colouring for prompts and the like
to use message_colour instead, so those "msg" abbreviations in 
food_colouring.txt actually have an effect now, and players can disable
colouring of prompts etc. (respects channels) while leaving it on in the 
inventory.
I'm leaving the menu_colour_prefix function in for now, in case it can
be used somewhere else.


git-svn-id: https://crawl-ref.svn.sourceforge.net/svnroot/crawl-ref/trunk@5558 c06c8d41-db1a-0410-9941-cceddc491573