Added several new tutorial triggers including information on being caught in a net, specific skills, magic resistance, and which stat to choose.
In a given game you'll only ever see a small fraction of all tutorial
events (we're at 71 now, and some of them are exclusive). In fact, a lot
of the information is so generally useful (esp. on gods, skills and items) 
that I'm seriously considering opening up the general information (How do I 
use item x? What does Spellcasting do?) to a wider selection of characters.
I don't think we'll want to explain the intricacies of Mummies, Ghouls,
or Vampires, but a lot of this stuff even applies to them... (If we
overhaul the tutorial, it's certainly not a 0.4 target, though.)

Resetting the tutorial version to 8 as we somehow switched from digital
version to binary and back to digital. It will take some time until we
actually reach 110, and by then, I am sure the base files won't be save
compatible anymore. :)


git-svn-id: https://crawl-ref.svn.sourceforge.net/svnroot/crawl-ref/trunk@5777 c06c8d41-db1a-0410-9941-cceddc491573