Apply patch 1952761 by jarpian (wasp): targetting with the monster list.
I've updated the patch to the current mlist code, but there are still a
couple of bugs that will need to be ironed out. For example, monsters
are picked by comparing their attitude and type, which is not sufficient
in the case of (non-collapsed) zombies, as well as mimics or ghosts.
Also if the number of monsters in the visible list is greater than its
rows (because some appear in the same row) then monsters outside of the
visible list can be targetted. I guess this is an unintended feature. :)
I'm certain there'll be more bugs, though.


git-svn-id: https://crawl-ref.svn.sourceforge.net/svnroot/crawl-ref/trunk@6129 c06c8d41-db1a-0410-9941-cceddc491573