Move around numerous restrictions in the secondary choices during character selection.
For weapons, I've applied the following rules:
* If there are significant differences between the aptitudes for
  the different starting weapons, only unrestrict the best ones.
* If all aptitudes are identical and <= 100, unrestrict all of them.
* If all aptitudes are identical but > 100, leave all of them
  restricted.
* If all aptitudes are really bad, but one of them is significantly
  better (say, 110 compared to 140) unrestrict that one.
* For non-Kobold gladiators, use spear restriction value for tridents.

For books, I've similarly used the aptitudes for Fire/Ice/Summonings.
I've tried to apply similar reasonings for Necromancy vs. Yredelemnul
(Invocations), but because Invocations is internally boosted few species
are even remotely as good in Necromancy as Invocations, so that doesn't
really work.
I don't understand the logic used to decide which species work good with
Xom as opposed to Makhleb.

The plan is to allow players to get an idea about the strengths and
weaknesses of their chosen character from the choices offered alone, and
to later be able to move these checks into functions of their own and then
use them for differentiating between random choices and good random
choices, like for species/class.


git-svn-id: https://crawl-ref.svn.sourceforge.net/svnroot/crawl-ref/trunk@6150 c06c8d41-db1a-0410-9941-cceddc491573