Fix the old issue of switching to the butchering tool before actually deciding on which corpse to butcher. Now, if there's no eligible tool to automatically switch to (or the autoswitch option is false) you get prompted at the beginning of the butchering routine but you don't actually switch weapons or take off your gloves until right before you start butchering. Obviously, the same applies when autoswitching. This means you don't lose a turn anymore if you change your mind in-between.
I really hope I haven't introduced other bugs this way but a fix was
badly in need since I'd changed the automatic weapon swap to take a turn
just as the manual swap does.

Also fix swap_when_safe so that it also working when the real weapon is 
on slot 'a' (== 0) which will probably be the norm rather than an
exception.


git-svn-id: https://crawl-ref.svn.sourceforge.net/svnroot/crawl-ref/trunk@6437 c06c8d41-db1a-0410-9941-cceddc491573