Fix 2014303: Vampire hunger properties Fix 2014275: Sif Muna giving manuals Fix 2013914: Draconians merging incorrectly
Also, at the start of a tutorial game change all secret doors within a
radius of 10 into closed doors, to avoid players encountering a dead-end
really close to the entry. (And yes, I've had that happen.)

Change display of forbidden choices in the character selection menus. It
should now be clearer what's going on.


git-svn-id: https://crawl-ref.svn.sourceforge.net/svnroot/crawl-ref/trunk@6463 c06c8d41-db1a-0410-9941-cceddc491573