Fix 2018458: Portal projectile not handling branded ammo or nets. Fix 2019803: cold/fire destroying all susceptible items on the floor Change "You don't eat raw eat!" to "Blech - you need greens!"
Also:
Experimentally change sacrifices to take only 1 turn again, no matter
how many corpses are actually sacrificed during this turn. Allow
chained 'yes' and 'no' answers during the butchering prompt, so that
'cccccc' is now really functionally identical to 'ca'.
In another return to previous handling, always get in that first turn of 
butchering for a single corpse (or the first corpse in butcher chains), 
so it takes a maximum of 3 turns rather than 4. Same for draining corpses 
as a Vampire: when draining corpses you get at least a little bit of 
nutrition out of it before you are interrupted.


git-svn-id: https://crawl-ref.svn.sourceforge.net/svnroot/crawl-ref/trunk@6579 c06c8d41-db1a-0410-9941-cceddc491573