First draft of the spell noisyness system: spells cast by the player and by wizard/priest monsters now make noise, with the loudness depending upon level and spell disciplines: conjuration spells have a loudness equal to their level, spells containing only air and/or poison equal to 50% of their level (rounded up), and everything else to 75% of the level (rounded up).  This undoudedtly needs rebalancing, and a better level+disciplines -> loudness formula.  Also, each spell can have it's loudness tweaked from the default via the noise_mod field of the spell_desc field in spl-data.h, for those spells whose desired loudness don't quite fit into whatever formula is used.
The messages used to announce a spell have been moved out to
source/dat/database/monspell.txt, which uses the same format and conventions as
monspeak.txt.  So far this has just been used to make invisible wizard/priest
monsters give a message to the player if the player can hear them cast, and
also to make curse skulls and Murray not "gesture wildly".  However, it could
be used to give a greater variety of spell announcement messages based upon the
spell used or the monster's type/species/genus, and/or to give multi-part
messages.


git-svn-id: https://crawl-ref.svn.sourceforge.net/svnroot/crawl-ref/trunk@7820 c06c8d41-db1a-0410-9941-cceddc491573