FR #2011386:
When giving or deleting a random mutation you can now request that the mutation
picked doesn't instantly kill the player.  Xom code now uses this option,
except for when the player is under penance from Xom or Xom is bored.

Improvements to the code that prevents Xom from accidently killing the player.


git-svn-id: https://crawl-ref.svn.sourceforge.net/svnroot/crawl-ref/trunk@7835 c06c8d41-db1a-0410-9941-cceddc491573