Also add variety to fixed level randarts, distinguishing between "easy" (1-3), "moderate" (4-6) and "difficult" (7-9).
Recount how often each spell school is used and use the two most
prominent schools (biasing by their order of definition in spl-util.h if
two are tied). 
This doesn't completely work yet: I got a book of Enchantment and Divination 
although not a single spell in the book belonged to Divination (5 of them were 
Enchantment, so that was okay). Also, the way spells are currently
selected (from two random disciplines) you can get a spell that does not fit 
either leading school (as later derived by counting) and might look odd with 
the title, e.g. a book is built on the premise of Conj/Fire and ends up
as Conj/Air with Fire/Trm thrown in.


git-svn-id: https://crawl-ref.svn.sourceforge.net/svnroot/crawl-ref/trunk@7848 c06c8d41-db1a-0410-9941-cceddc491573