Added "The Stairs" card (not yet in any deck), eventually to be the card which re-arranges all stairs on the level, but now is just used to test the code used to move stairs in LOS away or towards the player.
Added beam flavour BEAM_VISUAL for animating someting moving in a line which
doesn't have any effects on its own, and removed BEAM_LINE_OF_SIGHT, which was
no longer being used.  Added bolt structure field "delay", which can be used to
change the delay between animating each square in a beam from the default of 15
msec.

For eventual use as a Xom effect added the stairs-are-avoiding-you durations
DUR_REPEL_STAIRS_MOVE and DUR_REPEL_STAIRS_CLIMB.  If DUR_REPEL_STAIRS_MOVE
alone is set then stepping onto a stairs/portal/etc has a 50% chance of making
it move away in a random direction.  If DUR_REPEL_STAIRS_CLIMB alone is set
then attempting to climb a stairs/etc has a 50% chance of making it move away
before you can get through it (and ending your turn in the process).  If both
are set then moving onto it and attempting to climb each have a 29% chance of
making it move, for a combined chance of 49.59% of a move-and-climb making the
stairs move away.  Once a stair is successfully taken there's a 50% chance of
the stair on the other end moving away from the player, and then both durations
are reset to 0.

These can be tested by drawing The Stairs card ("&c stair"), which will set
DUR_REPEL_STAIRS_CLIMB if you're on top of a stair or DUR_REPEL_STAIRS_MOVE if
your'e not.

Added a few wizard targetting command placeholders.


git-svn-id: https://crawl-ref.svn.sourceforge.net/svnroot/crawl-ref/trunk@7865 c06c8d41-db1a-0410-9941-cceddc491573