FR #2083661: when polymorphed, unique (and named) monsters now retain their name (but not spells, yet), and use their old speech text if it makes sense for them to be able to do so.
Polymorphed monsters now retain some of their old flags, like
MF_CREATED_FRIENDLY and MF_GOT_HALF_XP.

If the Royal Jelly is polymorphed into something else then the Slime:6 vaults
will unlock when that monster is killed.

If the Royal Jelly is banished the player will be given a hint that it's gone
and that the vaults will remain locked.


git-svn-id: https://crawl-ref.svn.sourceforge.net/svnroot/crawl-ref/trunk@7970 c06c8d41-db1a-0410-9941-cceddc491573