Don't let chaos beam bounce off of one wall immediately into an adjacent wall; prevents chaos beams from bouncing *thorugh* walls.
If the beam passes out of bounds (the player was right at an edge of the map
and fired the beam off the map) then finish the beam and regress back into
bounds.

Always finish beam if it hits a wall and neither affects the wall nor bounces.

Regress out of walls before dropping an item.

Don't regress out of walls if target == source.

More assertions.


git-svn-id: https://crawl-ref.svn.sourceforge.net/svnroot/crawl-ref/trunk@7977 c06c8d41-db1a-0410-9941-cceddc491573