Make the spell message key suffix "gestures" works as documented.
If a targeted spell doesn't exactly hit anything (because the monster was
confused or the target was invisible) then, if appropriate, use the
prepositions "past" or "next to" instead of "at".

If a targeted spell with an invisible beam with a message saying "gestures at"
or "points to" doesn't exactly hit anything, then assume anything near the path
implied by the gesture/pointing is the desired target.


git-svn-id: https://crawl-ref.svn.sourceforge.net/svnroot/crawl-ref/trunk@8100 c06c8d41-db1a-0410-9941-cceddc491573