Remove the experience check for killing monsters with enslaved souls, since if you take their souls, you won't get any experience anyway.
This is also consistent with the behavior if you enslave a monster's 
soul, and then confuse it; if it e.g. steps in deep water and drowns, 
you'll get a normal kill with experience.  Seeing as it's difficult to 
manually move a newly created abomination or spectral thing to the 
nearest habitable ground if it's not there initially, and the potential 
oddities if there is no habitable ground nearby, I'm choosing to 
interpret the behavior as a feature.  As for an in-game explanation, you 
can only bind enslaved souls to you when you're near them, and if 
they're e.g. at the bottom of deep water, they're not near you.


git-svn-id: https://crawl-ref.svn.sourceforge.net/svnroot/crawl-ref/trunk@8312 c06c8d41-db1a-0410-9941-cceddc491573