Extend on Haran's monster resistance description. Thanks, Haran! :)
Now lists: 
* all resistances on one line, sorted by resistance level.
* susceptibility
* magic immunity
* see/sense invisible
* monster speed
* flying/levitating


An example output:

A fire vortex

A swirling cloud of flame.

It is extremely resistant to fire, very resistant to electricity, and
resistant to poison.
It is susceptible to cold.
It is immune to magical enchantments.
It is fast.
It can levitate.


I also removed all those symbol/quote/etc. checks from the database
lookup. Instead describe_monster() is now called directly, so the
descriptions are now identical when (v)iewing a monster and when
searching the database.

In other news, I removed Eringya's Surprising Bouquet since I doubt
we'll ever get it implemented, disagree balance-wise, and it was 
shamelessly nicked from Pratchett in the first place.
Breaks saves.


git-svn-id: https://crawl-ref.svn.sourceforge.net/svnroot/crawl-ref/trunk@8411 c06c8d41-db1a-0410-9941-cceddc491573