Implented some ranged brands from FR #2006917 and #1891231: shadow and penetration (not phasing) for launchers and shadow, penetration, dispersal, exploding, steel and silver for ammo.  Never randomly generated.
If a launcher of venom is used to launch flame or ice ammo then the resulting
bolt will be poisoned, just like poisoned ammo launched from a launcer of flame
or frost.

Put missile beam setup code that's common to monsters and the player in
setup_missile_beam().

Removed mons_thrown_object_destroyed(), thrown_object_destroyed() is now used
for both monsters and the player.

The bolt struct has several new callback fields that can be set to alter the
beam's behaviour; currently only used by the brands implemented in this commit,
but they should be general enough to be used by anything.

The bolt struct has the new field "special_explosion" which can be used to
cause an explosion with flavour and/or damage dice different than the rest of
the beam.


git-svn-id: https://crawl-ref.svn.sourceforge.net/svnroot/crawl-ref/trunk@8449 c06c8d41-db1a-0410-9941-cceddc491573